backend/requirements.txt
asgiref==3.3.1
autopep8==1.5.4
Django==3.1.6
django-cors-headers==3.7.0
django-rest-registration==0.5.6
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
django-filter==2.4.0
gunicorn
Pillow==8.1.2
psycopg2-binary==2.8.6
pycodestyle==2.6.0
PyJWT==2.0.1
pytz==2020.5
requests
sqlparse==0.4.1
toml==0.10.2
typing==3.7.4.3
